Add fixes to allow mesos with NFS jobFileStore runs.
1. Add a wait at worker-leader-worker interactions to allow the NFS
server to sync between clients before actions are taken. We observed
that workers were deleting their job  directories on completion and
returning before the NFS client has synced with the master. This way,
the master still sees the job directory and thinks the job still has
more to do but then errors out almost immediately.

2. Added a wait before deleting the fileJobStore so the clients can sync
before the deletion occures. We observed that shutil.rmtree errors on
jobstore delete claiming that a folder it is trying to delete within the
jobstore no longer exists. Also set this delete to ignore errors on
deletion.